+++ packages/moonstone/CHANGELOG.md
@@ -4,6 +4,11 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ## [unreleased]
 
+### Added
+
+- `moonstone/VideoPlayer` props: `backwardIcon`, `forwardIcon`, `jumpBackwardIcon`, `jumpForwardIcon`, `pauseIcon`, and `playIcon` to support icon customization of the player
+- `moonstone/VideoPlayer` props `jumpButtonsDisabled` and `rateButtonsDisabled` for disabling the pairs of buttons when it's inappropriate for the playing media
+
 ### Changed
 
 - `moonstone/Slider` colors and sizing to match the latest designs
@@ -13,7 +18,11 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Fixed
 
 - `moonstone/Input` to not spot its own input decorator on 5way out
+- `moonstone/VideoPlayer` to no longer render its `children` in multiple places
+- `moonstone/Input` to not spot its own input decorator when we 5way out.
 - `moonstone/Button` text color when used on a neutral (light) background in some cases
+- `moonstone/VideoPlayer` to no longer render its `children` in multiple places
+- `moonstone/Popup` background opacity
 
 ## [1.0.0] - 2017-03-31